数据处理和分析/requirements.txt
jieba-0.42.1
snownlp